Plaintext hover result: separate docs from type
In the plaintext hover result the docs come first, and nothing separates
it from the type.
This is visible when used through ALE in a terminal, where you'd get
something like this:
```
(**  [String.map f s] applies function [f] in turn to all the
    characters of [s] (in increasing index order) and stores the
    results in a new string that is returned.
    @since 4.00.0  *)(char -> char) -> string -> string
```

We could introduce a newline, or change the order to match the markdown
one (where doc starts with a newline if not empty).
I chose the latter.

Signed-off-by: Edwin Török <edwin@etorok.net>