Vendor dune libraries using submodules
use the nasty trick of creating the submodules in some random directory
and then copying the things that we need into the vendor directory.

This has the advantage that it makes it easy for us to keep up with
upstream. This will be important to use the rpc client library.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>