fix(ocamllsp): infer intf for closed docs
This commit fixes a bug when inferring the interface of a module whose
implementation hasn't been opened.

Since the module hasn't been opened by the user, we shouldn't open in
the document store and confuse the dynamic registration subscription
manager. In particular, if there is a promotion pending for the
document, this would wrongly remove the dynamic registration.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

ps-id: 70ff54c0-ab14-4ed5-82ed-e238173c0203