fix(ocamllex): do not raise exceptions on signatureHelp
As soon as I open a .mll file (with filetype set to ocamllex in NeoVim) the LSP
server sends a continuous stream "Document.dune" exceptions.
I currently have to turn off Lsp (using :LspStop) to be able to keep
editing.

Instead of raising an exception just return an empty result for
filetypes where signatureHelp is not implemented yet.

It'd probably be better to not advertise the capability, but not sure if
that is doable by filetype...

Signed-off-by: Edwin Török <edvin.torok@citrix.com>