fix: merlin diagnostics and watch mode
Updating the document with no changes is no longer needed because we
don't reuse the pipeline.

This also fixes a bug whenever we're editing a document and dune is
running in watch mode. The issue was that we were updating a document
the store in the background and overwriting updates to the document
requested by the user.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: a79143e1-01bc-4565-b900-ee5d97474ced -->