Refactor comment parsing to use odoc-parser and cmarkit
Omd 2.X doesn't support printing markdown, so OCaml LSP is blocked to Omd 1.X. This creates conflicts with codebases that have migrated to Omd 2.X (e.g. in ocamlorg).

To avoid conflicts for users of Omd 2.X, we migrate to cmarkit, which has a smaller dependency profile and supports printing to markdown.

Co-authored-by: Rafał Gwoździński <rafal.gwozdzinski@gmail.com>